P8_Construisez_et_testez_une_infrastructure_de_donnees_aymeric_bailleul/_projet/pyproject.toml
[tool.poetry]
name = "weather-etl-pipeline"
version = "1.0.0"
description = "Pipeline ETL pour données météorologiques - Lecture S3, nettoyage et préparation MongoDB"
authors = ["Aymeric Bailleul <aymericbailleul@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
boto3 = "^1.35.0"
pandas = "^2.2.0"
python-dotenv = "^1.0.0"
pymongo = "^4.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
